refactor: properly divide tracer into public and inner parts
