+++ src/main.rs
@@ -157,8 +157,7 @@ async fn main() -> color_eyre::Result<()> {
         .printer_from_cli(&tracing_args)
         .otel(otel_config)?
         .build_ptrace()?;
-      let tracer = Arc::new(tracer);
-      let tracer_thread = tracer.spawn(cmd, Some(output), token);
+      let (_tracer, tracer_thread) = tracer.spawn(cmd, Some(output), token)?;
       let mut errors = Vec::new();
       loop {
         match tracer_rx.recv().await {
@@ -241,9 +240,9 @@ async fn main() -> color_eyre::Result<()> {
         .printer_from_cli(&tracing_args)
         .otel(otel_config)?
         .build_ptrace()?;
-      let tracer = Arc::new(tracer);
 
       let frame_rate = tui_args.frame_rate.unwrap_or(60.);
+      let (tracer, tracer_thread) = tracer.spawn(cmd, None, token)?;
       let mut app = App::new(
         Some(PTracer {
           tracer: tracer.clone(),
@@ -255,7 +254,6 @@ async fn main() -> color_eyre::Result<()> {
         baseline,
         pty_master,
       )?;
-      let tracer_thread = tracer.spawn(cmd, None, token);
       let mut tui = tui::Tui::new()?.frame_rate(frame_rate);
       tui.enter(tracer_rx)?;
       app.run(&mut tui).await?;
@@ -309,8 +307,7 @@ async fn main() -> color_eyre::Result<()> {
         .printer_from_cli(&tracing_args)
         .otel(otel_config)?
         .build_ptrace()?;
-      let tracer = Arc::new(tracer);
-      let tracer_thread = tracer.spawn(cmd, None, token);
+      let (_tracer, tracer_thread) = tracer.spawn(cmd, None, token)?;
       let mut errors = Vec::new();
       match format {
         ExportFormat::OpenTelemetry => todo!(),
